Move `aio_*` functions to `base/aio.cpp/h`
And move `ASYNCIO` type to `base/types.h`.
