+++ src/base/log.cpp
@@ -1,3 +1,4 @@
+#include "aio.h"
 #include "color.h"
 #include "logger.h"
 #include "system.h"
